@@ -3,9 +3,11 @@
   "version": "1.0.0",
   "description": "A 1.1.1.1 DNS resolver built for Discord",
   "main": "dist/worker.js",
+  "type": "module",
   "scripts": {
-    "build": "webpack",
+    "build": "NODE_OPTIONS=--experimental-json-modules webpack",
     "dev": "NODE_ENV=development wrangler dev",
+    "tunnel": "cloudflared tunnel --url http://localhost:8787",
     "publish:production": "NODE_ENV=production wrangler publish -e production",
     "publish:staging": "NODE_ENV=staging wrangler publish -e staging",
     "logs:production": "wrangler tail -e production",
@@ -25,21 +27,22 @@
   },
   "homepage": "https://github.com/MattIPv4/DNS-over-Discord#readme",
   "dependencies": {
-    "discord-api-types": "^0.24.0",
-    "is-valid-domain": "^0.1.2",
+    "discord-api-types": "^0.31.1",
+    "is-valid-domain": "^0.1.6",
     "web-whois": "0.0.6",
     "workers-sentry": "0.0.6"
   },
   "devDependencies": {
-    "@babel/core": "^7.15.8",
-    "@babel/plugin-proposal-optional-chaining": "^7.14.5",
-    "@cloudflare/wrangler": "^1.19.3",
-    "babel-loader": "^8.2.2",
+    "@babel/core": "^7.17.9",
+    "@babel/eslint-parser": "^7.17.0",
+    "@babel/plugin-syntax-import-assertions": "^7.16.7",
+    "@cloudflare/wrangler": "^1.19.11",
     "deep-equal": "^2.0.5",
-    "dotenv": "^10.0.0",
-    "eslint": "^8.0.1",
-    "node-fetch": "^2.6.7",
-    "webpack": "^5.58.2",
-    "webpack-cli": "^4.9.0"
+    "dotenv": "^16.0.0",
+    "eslint": "^8.13.0",
+    "eslint-plugin-import": "^2.26.0",
+    "node-fetch": "^3.2.3",
+    "webpack": "^5.72.0",
+    "webpack-cli": "^4.9.2"
   }
 }
