@@ -1,12 +1,12 @@
-const { InteractionResponseType, ApplicationCommandOptionType, ComponentType } = require('discord-api-types/payloads/v9');
-const { VALID_TYPES } = require('../utils/dns');
-const { validateDomain, handleDig } = require('../utils/dig');
-const { editDeferred } = require('../utils/discord');
-const { component } = require('../components/dig-refresh');
+import { InteractionResponseType, ApplicationCommandOptionType, ComponentType } from 'discord-api-types/payloads/v9';
+import { VALID_TYPES } from '../utils/dns.js';
+import { validateDomain, handleDig } from '../utils/dig.js';
+import { editDeferred } from '../utils/discord.js';
+import digRefresh from '../components/dig-refresh.js';
 
 const optionTypes = Object.freeze(VALID_TYPES.slice(0, 25)); // Discord has a limit of 25 options
 
-module.exports = {
+export default {
     name: 'dig',
     description: 'Perform a DNS over Discord lookup',
     options: [
@@ -58,7 +58,7 @@ module.exports = {
                 components: [
                     {
                         type: ComponentType.ActionRow,
-                        components: [ component ],
+                        components: [ digRefresh.component ],
                     },
                 ],
             });
