@@ -1,7 +1,7 @@
-const { InteractionResponseType } = require('discord-api-types/payloads/v9');
-const { createEmbed } = require('../utils/embed');
+import { InteractionResponseType } from 'discord-api-types/payloads/v9';
+import { createEmbed } from '../utils/embed.js';
 
-module.exports = {
+export default {
     name: 'github',
     description: 'Get a link to the open-source GitHub repository for DNS over Discord',
     execute: async ({ response }) => response({
