@@ -1,18 +1,20 @@
-const path = require('path');
+import { fileURLToPath } from 'url';
 const NODE_ENV = process.env.NODE_ENV || 'development';
-const env = require('dotenv').config({ path: path.join(__dirname, `${NODE_ENV}.env`) });
-const { DefinePlugin } = require('webpack');
-const WorkersSentryWebpackPlugin = require('workers-sentry/webpack');
-const build = require('./src/build');
+import dotenv from 'dotenv';
+const env = dotenv.config({ path: fileURLToPath(new URL(`${NODE_ENV}.env`, import.meta.url)) });
+
+import webpack from 'webpack';
+import WorkersSentryWebpackPlugin from 'workers-sentry/webpack.js';
+import build from './src/build/index.js';
 
 console.log(`Using ${NODE_ENV} environment for build...`);
 
-module.exports = {
+export default {
     mode: 'none',
     target: 'webworker',
     entry: './src/index.js',
     output: {
-        path: path.join(__dirname, 'dist'),
+        path: fileURLToPath(new URL('dist', import.meta.url)),
         filename: 'worker.js',
     },
     plugins: [
@@ -20,10 +22,13 @@ module.exports = {
         { apply: compiler => compiler.hooks.beforeRun.tapPromise('PrepareBuildBeforeWebpack', build) },
 
         // Expose our environment in the worker
-        new DefinePlugin(Object.entries(env.parsed).reduce((obj, [ key, val ]) => {
+        new webpack.DefinePlugin(Object.entries(env.parsed).reduce((obj, [ key, val ]) => {
             obj[`process.env.${key}`] = JSON.stringify(val);
             return obj;
         }, { 'process.env.NODE_ENV': JSON.stringify(NODE_ENV) })),
+
+        // Ensure single chunk
+        new webpack.optimize.LimitChunkCountPlugin({ maxChunks: 1 }),
 
         // Publish source maps to Sentry on each build
         new WorkersSentryWebpackPlugin(
@@ -32,11 +37,8 @@ module.exports = {
             process.env.SENTRY_PROJECT,
         ),
     ],
-    module: {
-        rules: [
-            {
-                test: /\.js$/, exclude: /node_modules/, loader: 'babel-loader',
-            },
-        ],
+    externals: {
+        // Don't webpack node-fetch, rely on fetch global
+        'node-fetch': 'fetch',
     },
 };
